Uploading images
This is a much bigger commit than it should be, and honestly, it's still not a full vertical slice. A lot of the blame here goes to me not really knowing how to write unit tests around thumbnails. If I did, I'd probably be able to make suitably smaller commits. But I want to commit this before I go into fixing up the Form component with a bunch of new, smaller components, which will then allow me to circle back and make sure that everything here actually works. This does have all of the front-end pieces for drag-and-drop image upload and selecting a thumbnail. Actually uploading that is the part that's still in question.