Don't wrap
The GeoJSON doesn't wrap, so letting the tile layer wrap can create some confusing situations. It would be great if we could limit panning so that you can't pan off the edge of the world into paradoxical greyness, but that isn't something that really needs to be done right now to launch the site, so we'll see if that's something we can fix later.