Fix redirect bug
Because we started handling stuff client-side for form submission, we're stuck with testing if we're supposed to move forward to the page created or updated, or if there was an error and we should stay here. We were testing `res.config.url`, but that's not the right place to look. We should be looking at `res.request.responseURL`.