Is it coastal?
Separate method to load sea level GeoJSON files into a separate method, because it's now used in two places: when loading the map and also when creating a community. We also add a new geo util that creates a GeoJSON polygon approximating a circle, based on a particular point. We can use that to create a circle centered on the coordinates given with a 45km radius, cycle through the sea level GeoJSON features, and recognize the territory as coastal if the circle intersects any of the other features (e.g., if we encounter the sea within 45km of the center in any direction).