Don't track steps
Explicitly tracking steps in the state means that we have a "magic number" that will make things more brittle as time goes on. We can figure out the next step based on what information we have in state and what's still missing. This puts all of the logic for ordering steps in one place (the CommunityCreation component's new renderStep method).