Using the new db structure
THIS IS A DANGEROUSLY LARGE COMMIT. I AM TERRIFIED.

Lots of test updates to handle foreign keys (the order in which we clean up the database matters now), and lots of tweaks to the Page model to use the new database structure. Seems to work locally, but with any commit this big you're really rolling the dice.