checkUntil
There will be times when we want to roll on a random table, but we'll want to roll again if we get a particular result (ex: we're aging up a marriage candidate; obviously, this isn't someone who's going to die, so death is not an acceptable outcome). checkUntil keeps checking until it gets a result that is not unacceptable. If every possibility is unacceptable, it just returns false.