Choose sex and gender
I was going to add gender as characters age, but that's dumb. That might be how people work in the real world, but these aren't real people, they're just a statistical model. There's no need for that added complexity. But we *do* need to be able to say, "Hey, I need a random woman," or "I need a random man," or whatever when we get to populating potential relationship partners, so we do have the ability to provide a gender and get a sex, or just get them both randomly.