Big refactor
It's hard to test random stuff, and the way I'd written all this before, those tendrils of randomness were spreading and making huge chunks of code untestable. So this is a big refactor, moving all that code into two classes — Community and Person — which makes it all a lot easier to test. There's 77 new unit tests in here!