Generating a random partner
We generate a random partner and pick one. Later, we'll need to handle how these are tagged in the database so that when some other community is being created, and this character might've been born there, we can insert her. We'll also need to later add a database query for all the people who left their communities at that time to join a partner in the right area, and add them to the list, mixed in with the random characters.  But that's for later. We need to get more of the basics in place before we worry about these connections between iterations.