Fixes to finding a partner
Firstly, we were running the dating app from Hell here. The way personality distance is calculated, the higher the number, the less you have in common. And then that random love factor is subtracted. So the lower the love score, the more compatible two people are. But then we were picking the MAX, as in, the person you had the least in common with and felt the least for! Yikes! Switched that to min, as it should be.

Secondly, added a threshold. If you don't find someone with whom you at least have an average connection, then sorry, you just didn't find anyone.