Mo problems
Added a "yield" varaible to the territory. Each year it regenerates the community's carrying capacity (+150 for a village or +30 for a hunter-gatherer band), and then it subtracts the current population. If you're below your carrying capacity, then you'll begin accumulating. If you're over, you'll start dropping, and can go negative. This should create the sort of turbulent movement towards the carrying capacity seen in ecological models.

The method for deciding when problems develop uses this. If yield is negative, you're in lean times, and if it's not, you're not. That one's simple. Being in lean times becomes a factor for sickness and conflict, though. Sickness also adds in population, with your population relative to your carrying capacity adding another factor for the likelihood of sickness.