Polycule class
I always intended to write a Polycule class. Originally, I'd thought that it would be a class for containing a set of Pair instances. But now I see that's insufficient, so this Polycule class will have to replace the Pair class entirely.

We'll still have monogamous couples. While some 86% of the world's cultures are polygamous, as a matter of practical fact most of the people in polygamous societies are in monogamous relationships, not because monogamy is an enforced cultural norm, but simply because it's too expensive or difficult to introduce any more partners than that. I'll be aiming for that with the numbers in the Polycule class. But monogamous couples are just a special case for the Polycule class, where the number of people happens to be two.