Add in factors for injury and illness
The more open you are to new experiences, the more chances you have to get hurt or sick. In times of conflict, the odds of getting hurt are doubled. In times of sickness, the odds of getting sick are quadrupled, and in times of scarcity they're doubled (making you 8x more likely to get sick if the community is suffering both scarcity and sickness, which is entirely possible since scarcity makes sickness more likely).