The community will need a list of polycules to iterate over
When we run a simulation of each year, first we'll change the polycules, because indiividuals use that information when they're aging. For example, someone dumped might try to start a new relationship, and anyone in a polycule that's having a baby doesn't have a lot of time left over for mastering new skills.