How do we add founders?
Last commit of 2019! And it's an appropriate one. This one shifts us from "getting our basic pieces together" to "how do we fit this all together." We're now in the part where I need to analyze how it all comes together, tweak stuff, and makes sure it comes out in a sensible way.