Effects of conflict, sickness, and lean times
A conflict means you have a second chance of getting hurt, and especially high chances of getting hurt if you're between 16 and 34.

Sickness going around makes it 10x more likely that you'll get sick, and lean times makes it 2x more likely that you'll get sick.

These are all tested individually, so it's possible (though unlikely) to suffer an injury from dumb luck, and then suffer an injury in a fight, and then get sick just from being unlucky, and then catching whatever is going around, and then getting sick a third time because it's lean times and you don't have enough to eat. That should be exceptionally rare, but it's sure to happen to some woesome bastard some day.