Person.feelsSafe
Does this person feel safe and secure enough to undertake a major life change, like starting a relationship or having a baby? The less open you are to new experience you are, the more cautious and conservative you are, so the more years of peace and abundance you need in the community's recent history to feel safe.