Let's start again, please
Ran into a really basic problem with how I was doing this: circular references make things impossible to clone. The good news is (1) I should be able to rebuild a lot, because I can copy over most of the trickiest stuff, and (2) I've learned a lot about how to do this from the previous two attempts.