Body.constructor: Determine sex
This is in the Body class, so we're concerned here with biological sex, and in particular one's capacity to sexually reproduce. "Male" and "female" are useful shorthands in this context for "is physically capable of contributing sperm to produce a viable offspring" and "is physically capable of contributing an egg, gestating, and giving birth to a viable offspring," respectively.