Skills.pick: Factoring in rare, discouraged, and younger skills.
Rare skills are those where you kind of need someone to teach you, and those are in short supply. Discouraged skills are those that the community does not approve of (like deception and sorcery). Some skills are marked as younger (like running), where it's harder to learn when you're older. If you pick a skill that you're a little old for, or one where you need a teacher but it's hard to find one, pick again. If you pick it again, OK, go ahead and start learning it. If you pick a discouraged skill, you can learn it if you fail an agreeableness check (you just dgaf), or if you pick it a second time (you might gaf, but you just really, really want to do this).