Person.separation: Call them "separated" not "parties"
Reports can layer in chain reactions. For example, when an act of adultery leads to separation, we'll have a list of ID's called `adulterers` for those who were part of the act of adultery, as well as a separate list of ID's produced by this function. Calling that list `parties` makes it a bit more ambiguous — parties to the adultery, or to the separation? Changing the name to `separated` helps alleviate that ambiguity.