Community.adjustMonogamy
Monogamy is a pretty strong norm in the modern world, but by the time of the Fifth World, we'll likely return to the more common state of affairs, where polygamy is accepted in 85% or so of societies, even though, practically speaking, most people still wind up with just one partner, simply because that's how things are going to work out with a roughly 50/50 gender split. We want to be able to trace the decline of this institution, though, with people early on who are reticent to challenge social norms, and it falling to those most willing to buck social trends to innovate and change society. We'll be aiming for monogamy to reach zero in about 85% of all simulations, but this is the core of how that happens. Each person in the community who just lives her life in a polygamous relationship erodes monogamy as a norm, because polygamy isn't a weird, scary thing, it's just Aunt Jessica. But each adultery that rocks the community and involves a polygamous individual is going to reinforce those norms, convincing everyone that this is why we need monogamy, to avoid exactly this kind of drama.